Moved getCurrentPosition logic to separate module
